// src/shared/api/schema.d.ts
export interface paths {
// ...
    "/api/v2/posts/list/{boardId}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * 비회원용 게시판 글 조회 API
         * @description 게시판에서 글을 조회하는 API 입니다.
         */
        get: operations["getBoardById"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
        /**
         * @description 차단 유형
         * @enum {string}
         */
        BanType: "NONE" | "WARNING" | "BAN_1D" | "BAN_3D" | "BAN_5D" | "BAN_1W" | "BAN_2W" | "BAN_1M" | "PERMANENT";
        ReportProcessRequest: {
            banType: components["schemas"]["BanType"];
            processReason?: string;
        };
        ApiResponseReportProcessResponse: {
            /** Format: int32 */
            status: number;
            message: string;
            code?: string;
            data?: components["schemas"]["ReportProcessResponse"];
        };
// ...
        PositionRequest: {
            mainP: components["schemas"]["Position"];
            subP: components["schemas"]["Position"];
            wantP: ("ANY" | "TOP" | "JUNGLE" | "MID" | "ADC" | "SUP")[];
        };
        IsMikeRequest: {
            mike: components["schemas"]["Mike"];
        };
// ...
        MyProfileResponse: {
            /** Format: int64 */
            id: number;
            /** Format: int32 */
            profileImg: number;
            mike: components["schemas"]["Mike"];
            email: string;
            gameName: string;
            tag: string;
            soloTier: components["schemas"]["Tier"];
// ...
        };
        /**
         * @description 리그오브레전드 티어
         * @enum {string}
         */
        Tier: "UNRANKED" | "IRON" | "BRONZE" | "SILVER" | "GOLD" | "PLATINUM" | "EMERALD" | "DIAMOND" | "MASTER" | "GRANDMASTER" | "CHALLENGER";
        BoardUpdateRequest: {
            gameMode: components["schemas"]["GameMode"];
            mainP: components["schemas"]["Position"];
            subP: components["schemas"]["Position"];
            wantP: ("ANY" | "TOP" | "JUNGLE" | "MID" | "ADC" | "SUP")[];
            mike?: components["schemas"]["Mike"];
            /** @description 게임 스타일 리스트 (선택, 최대 3개) */
            gameStyles?: number[];
            /** @description 게시글 내용 (선택) */
            contents?: string;
        };
// ...
        ReportListResponse: {
            /** Format: int64 */
            reportId: number;
            /** Format: int64 */
            fromMemberId?: number;
            fromMemberName: string;
            fromMemberTag: string;
            /** Format: int64 */
            toMemberId: number;
            toMemberName: string;
            toMemberTag: string;
            content?: string;
// ...
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
    processReport: {
        parameters: {
            query?: never;
            header?: never;
            path: {
                /** @description 처리할 신고의 ID입니다. */
                reportId: number;
            };
            cookie?: never;
        };
        requestBody: {
            content: {
                "application/json": components["schemas"]["ReportProcessRequest"];
            };
        };
        responses: {
            /** @description OK */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "*/*": components["schemas"]["ApiResponseReportProcessResponse"];
                };
            };
            /** @description [REPORT_402] 해당 신고 건을 찾을 수 없습니다. */
            404: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ApiErrorResponse"];
                };
            };
        };
    };
// ...
    boardList: {
        parameters: {
            query: {
                page: number;
                /** @description (선택) 게임 모드를 입력해주세요. < 빠른대전: FAST, 솔로랭크: SOLO, 자유랭크: FREE, 칼바람 나락: ARAM > */
                gameMode?: components["schemas"]["GameMode"];
                /** @description (선택) 티어를 선택해주세요. */
                tier?: components["schemas"]["Tier"];
                /** @description (선택) 주 포지션을 입력해주세요. < 전체: ANY, 탑: TOP, 정글: JUNGLE, 미드: MID, 원딜: ADC, 서포터: SUP > */
                mainP?: components["schemas"]["Position"];
                /** @description (선택) 부 포지션을 입력해주세요. < 전체: ANY, 탑: TOP, 정글: JUNGLE, 미드: MID, 원딜: ADC, 서포터: SUP > */
                subP?: components["schemas"]["Position"];
                /** @description (선택) 마이크 여부를 선택해주세요. */
                mike?: components["schemas"]["Mike"];
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
// ...
    };
// ...
}